Move NamespaceGroup class and its helper function to a new file
